Bumped version, moved checker to separate file
