+++ api/services/Flairs.js
@@ -1,3 +1,5 @@
+var sha1 = require('node-sha1');
+
 exports.formattedName = function(name) {
   if (!name) {
     return "";
@@ -24,4 +26,21 @@ exports.formattedName = function(name) {
     formatted += suffix;
   }
   return formatted;
+};
+
+exports.isValid = function(code) {
+  code = code.replace(/-/g,'');
+  if (!code.match(/^\d{12}$/) || code > 549755813887) {
+      return 0;
+  }
+  var checksum = Math.floor(code/4294967296);
+  var byte_seq = (code % 4294967296).toString(16);
+  while (byte_seq.length < 8) { byte_seq = "0"+byte_seq; }
+  var byte_arr = byte_seq.match(/../g).reverse();
+  var hash_seq = "";
+  for (var i = 0; i < 4; i++) {
+      hash_seq += String.fromCharCode(parseInt(byte_arr[i],16));
+  }
+  var new_chk = (parseInt(sha1(hash_seq).substring(0,2),16) >> 1);
+  return (new_chk == checksum)?1:0;
 };
