+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "reddit-flair-apps",
   "private": true,
-  "version": "1.5.8",
+  "version": "1.5.9",
   "description": "A small project to allow easy adding of flair applications for subreddits (focusing initially on /r/pokemontrades) and easy moderation for moderators.",
   "keywords": [],
   "dependencies": {
